Now writers have double4 pos array.

+++ src/VTKLegacyWriter.h
@@ -36,7 +36,7 @@ public:
 	VTKLegacyWriter(const Problem *problem);
 	~VTKLegacyWriter();
 
-	void write(uint numParts, const float4 *pos, const float4 *vel,
+	void write(uint numParts, const double4 *pos, const float4 *vel,
 			const particleinfo *info, const float3 *vort, float t,const bool testpoints, const float4 *normals);
 };
